+++ packages/sui-hoc/src/withIntersectionObserver.js
@@ -4,7 +4,8 @@ import React, {Component} from 'react'
 const shouldLoadIntersectionObserver = () =>
   !('IntersectionObserver' in window) ||
   !('IntersectionObserverEntry' in window) ||
-  !('intersectionRatio' in window.IntersectionObserverEntry.prototype)
+  !('intersectionRatio' in window.IntersectionObserverEntry.prototype) ||
+  !('isIntersecting' in window.IntersectionObserverEntry.prototype)
 
 export const hocIntersectionObserverWithOptions = (
   options = {}
@@ -44,7 +45,9 @@ export const hocIntersectionObserverWithOptions = (
     }
 
     componentWillUnmount() {
-      this.state.intersectionObserver.disconnect()
+      if (this.state.intersectionObserver) {
+        this.state.intersectionObserver.disconnect()
+      }
     }
 
     render() {