+++ packages/sui-mono/bin/sui-mono-phoenix.js
@@ -2,10 +2,8 @@
 /* eslint no-console:0 */
 const program = require('commander')
 const {getSpawnPromise, parallelSpawn, showError} = require('@s-ui/helpers/cli')
-const {existsSync} = require('fs')
-const path = require('path')
 
-const config = require('../src/config')
+const {checkIsMonoPackage, getWorkspaces} = require('../src/config')
 
 const DEFAULT_CHUNK = 5
 
@@ -119,12 +117,10 @@ const installRootPackages = () => {
 }
 
 const executePhoenixOnPackages = () => {
-  if (config.isMonoPackage()) return
+  if (checkIsMonoPackage()) return
 
   // get scopes only where a `npm install` is possible
-  const scopes = config
-    .getScopesPaths(scopeArgument)
-    .filter(folder => existsSync(path.resolve(folder, 'package.json')))
+  const scopes = getWorkspaces()
 
   const removePackagesCommands = scopes.map(cwd => [...RIMRAF_CMD, {cwd}])
   const installPackagesCommands = scopes.map(createInstallPackagesCommand)