+++ packages/sui-mono/bin/sui-mono-release.js
@@ -2,7 +2,11 @@
 
 const program = require('commander')
 const path = require('path')
-const config = require('../src/config')
+const {
+  checkIsMonoPackage,
+  getChangelogFilename,
+  getPublishAccess
+} = require('../src/config')
 const checker = require('../src/check')
 const {serialSpawn, showError} = require('@s-ui/helpers/cli')
 const {getPackageJson} = require('@s-ui/helpers/packages')
@@ -42,10 +46,10 @@ program
   .parse(process.argv)
 
 const BASE_DIR = process.cwd()
-const packagesFolder = config.getPackagesFolder()
-const publishAccess = config.getPublishAccess()
+
+const publishAccess = getPublishAccess()
 const suiMonoBinPath = require.resolve('@s-ui/mono/bin/sui-mono')
-const changelogFilename = config.getChangelogFilename()
+const changelogFilename = getChangelogFilename()
 
 const RELEASE_CODES = {
   0: 'clean',
@@ -59,53 +63,46 @@ const scopeMapper = ({scope, status}) => ({
   code: status[scope].increment
 })
 
-const releasesByPackages = ({status}) =>
-  Object.keys(status).map(scope => scopeMapper({scope, status}))
-
-const singlePackageRelease = ({status, packageScope}) =>
-  Object.keys(status)
-    .filter(scope => scope === packageScope)
+const releasesByPackages = ({status}) => {
+  const {scope: packageScope} = program
+  return Object.keys(status)
+    .filter(scope => (packageScope ? scope === packageScope : true))
     .map(scope => scopeMapper({scope, status}))
+}
 
-const releaseEachPkg = ({pkg, code} = {}, {skipCI} = {}) => {
+const releaseEachPkg = ({pkg, code, skipCI} = {}) => {
   return new Promise((resolve, reject) => {
     if (code === 0) return resolve()
 
-    const isMonoPackage = config.isMonoPackage()
-
+    const isMonoPackage = checkIsMonoPackage()
     const tagPrefix = isMonoPackage ? '' : `${pkg}-`
-
     const packageScope = isMonoPackage ? 'META' : pkg.replace(path.sep, '/')
 
-    const cwd = isMonoPackage
-      ? BASE_DIR
-      : path.join(BASE_DIR, packagesFolder, pkg)
-    const pkgInfo = require(path.join(cwd, 'package.json'))
-    const scripts = pkgInfo.scripts || {}
+    const cwd = isMonoPackage ? BASE_DIR : path.join(process.cwd(), pkg)
+    const {private: isPrivatePackage} = getPackageJson(cwd, true)
 
     const releaseCommands = [
       ['npm', ['--no-git-tag-version', 'version', `${RELEASE_CODES[code]}`]],
       ['git', ['add', path.join(cwd, 'package.json')]]
     ]
+
     const docCommands = [
       [suiMonoBinPath, ['changelog', cwd]],
       ['git', ['add', path.join(cwd, changelogFilename)]],
       ['git', ['commit --amend --no-verify --no-edit']]
     ]
+
     const publishCommands = [
-      scripts.build && ['npm', ['run', 'build']],
-      !pkgInfo.private && ['npm', ['publish', `--access=${publishAccess}`]],
+      !isPrivatePackage && ['npm', ['publish', `--access=${publishAccess}`]],
       ['git', ['push', '-f', '--tags', 'origin', 'HEAD']]
     ].filter(Boolean)
 
     serialSpawn(releaseCommands, {cwd})
       .then(() => {
-        // Create release commit
         const {version} = getPackageJson(cwd, true)
 
-        // We're adding [skip ci] to the commit message to avoid
-        // start a build on CI if not needed
-        // docs: https://docs.travis-ci.com/user/customizing-the-build/#skipping-a-build
+        // Add [skip ci] to the commit message to avoid CI build
+        // https://docs.travis-ci.com/user/customizing-the-build/#skipping-a-build
         const commitMsg = `release(${packageScope}): v${version}${
           skipCI ? ' [skip ci]' : ''
         }`
@@ -127,30 +124,15 @@ const releaseEachPkg = ({pkg, code} = {}, {skipCI} = {}) => {
   })
 }
 
-const packageScope = program.scope
-const releaseMode = packageScope ? singlePackageRelease : releasesByPackages
+const checkIsMasterBranchActive = async ({cwd}) => {
+  const {stdout} = await exec(`git rev-parse --abbrev-ref HEAD`, {
+    cwd
+  })
 
-const checkIsMasterBranchActive = async ({status, cwd}) => {
-  try {
-    const {stdout} = await exec(`git rev-parse --abbrev-ref HEAD`, {
-      cwd
-    })
-
-    if (stdout.trim() === 'master') {
-      return Promise.resolve(status)
-    } else {
-      throw new Error(
-        'Active branch is not master, please make releases only in master branch'
-      )
-    }
-  } catch (error) {
-    showError(error)
-
-    return Promise.reject(error)
-  }
+  return stdout.trim() === 'master'
 }
 
-const automaticRelease = async ({
+const prepareAutomaticRelease = async ({
   githubToken,
   githubUser,
   githubEmail,
@@ -177,41 +159,42 @@ const automaticRelease = async ({
   await exec(`git pull origin master`, {cwd})
 }
 
-const isAutomaticRelease = ({githubToken, githubUser, githubEmail}) => {
-  return githubToken && githubUser && githubEmail
-}
+const checkIsAutomaticRelease = ({githubToken, githubUser, githubEmail}) =>
+  githubToken && githubUser && githubEmail
 
-checker
-  .check()
-  .then(async status => {
-    const {githubEmail, githubToken, githubUser} = program
-    isAutomaticRelease({
-      githubEmail,
-      githubToken,
-      githubUser
-    })
-      ? await automaticRelease({
+checkIsMasterBranchActive({cwd: process.cwd()})
+  .then(isMaster => {
+    if (!isMaster) {
+      console.warn('Active branch is not master branch. No releases to do.')
+      return
+    }
+
+    return checker.check().then(async status => {
+      const {githubEmail, githubToken, githubUser} = program
+
+      if (checkIsAutomaticRelease(program)) {
+        await prepareAutomaticRelease({
           githubEmail,
           githubToken,
           githubUser,
           cwd: process.cwd()
         })
-      : await checkIsMasterBranchActive({status, cwd: process.cwd()})
-    return status
-  })
-  .then(status => releaseMode({status, packageScope}))
-  .then(releases =>
-    releases
-      .filter(({code}) => code !== 0)
-      .map(release => () => releaseEachPkg(release, {skipCI: program.skipCi}))
-      // https://gist.github.com/istarkov/a42b3bd1f2a9da393554
-      .reduce(
-        (m, p) => m.then(v => Promise.all([...v, p()])),
-        Promise.resolve([])
+      }
+
+      return (
+        releasesByPackages({status})
+          .filter(({code}) => code !== 0)
+          .map(release => () =>
+            releaseEachPkg({...release, skipCI: program.skipCi})
+          )
+          // https://gist.github.com/istarkov/a42b3bd1f2a9da393554
+          .reduce(
+            (m, p) => m.then(v => Promise.all([...v, p()])),
+            Promise.resolve([])
+          )
       )
-  )
+    })
+  })
   .catch(err => {
-    console.error('[sui-mono release] ERROR:')
-    console.error(err)
-    process.exit(1)
+    showError(`[sui-mono release]: ${err}`)
   })