+++ packages/sui-pde/src/adapters/default.js
@@ -1,0 +1,17 @@
+export default class DefaultAdapter {
+  getEnabledFeatures({attributes}) {
+    return Promise.resolve([])
+  }
+
+  getInitialData() {
+    return null
+  }
+
+  activateExperiment() {
+    return null
+  }
+
+  getVariation() {
+    return null
+  }
+}