+++ packages/sui-pde/src/index.js
@@ -1,0 +1,4 @@
+export {default as PDE} from './pde'
+export {default as useFeature} from './hooks/useFeature'
+export {default as PdeContext} from './contexts/PdeContext'
+export {default as useExperiment} from './hooks/useExperiment'