+++ packages/sui-ssr/server/hooksFactory/index.js
@@ -3,9 +3,7 @@ import {readFile} from 'fs'
 import {promisify} from 'util'
 import {resolve} from 'path'
 import {getTplParts, HtmlBuilder} from '../template'
-import {publicFolderByHost, hrTimeToMs} from '../utils'
-import ssrConf from '../config'
-
+import {publicFolder, hrTimeToMs, siteByHost} from '../utils'
 import {createServerContextFactoryParams} from '@s-ui/react-initial-props'
 
 // __MAGIC IMPORTS__
@@ -35,11 +33,7 @@ const getStaticErrorPageContent = async (status, req) => {
     return __PAGES__[status]
   }
   const html = await promisify(readFile)(
-    resolve(
-      process.cwd(),
-      ssrConf.multiSite ? publicFolderByHost(req) : 'public',
-      `${status}.html`
-    ),
+    resolve(process.cwd(), publicFolder(req), `${status}.html`),
     'utf8'
   ).catch(() => `Generic Error Page: ${status}`)
   __PAGES__[status] = html
@@ -81,6 +75,8 @@ export const hooksFactory = async () => {
     [TYPES.ROUTE_MATCHING]: async (req, res, next) => {
       const startRouteMatchingTime = process.hrtime()
       const {performance = {}, url} = req
+      const site = siteByHost(req)
+      const siteRoutes = site && isFunction(routes) ? routes({site}) : routes
 
       match[promisify.custom] = args =>
         new Promise((resolve, reject) => {
@@ -93,14 +89,14 @@ export const hooksFactory = async () => {
           })
         })
 
-      if (!routes) {
+      if (!siteRoutes) {
         console.error('Required routes file missing') // eslint-disable-line no-console
         process.exit(1)
       }
 
       try {
         const {redirectLocation, renderProps} = await promisify(match)({
-          routes,
+          routes: siteRoutes,
           location: url
         })