+++ packages/sui-studio/src/environment-mocha/setupEnvironment.js
@@ -1,16 +1,16 @@
 import SUIContext from '@s-ui/react-context'
 import {render} from '@testing-library/react'
 
-const setupEnvironment = (Component, context) => props => {
+const setupEnvironment = (Component, context) => (props, ref) => {
   const container = document.createElement('div')
   container.setAttribute('id', 'test-container')
   return render(
     context ? (
       <SUIContext.Provider value={context}>
-        <Component {...props} />
+        <Component {...props} ref={ref} />
       </SUIContext.Provider>
     ) : (
-      <Component {...props} />
+      <Component {...props} ref={ref} />
     ),
     {
       container: document.body.appendChild(container)